pyproject.toml
[tool.poetry]
name = "comments"
version = "0.1.0"
description = ""
authors = ["Evhenii Podolian <alfan2620@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
drf-yasg = "^1.21.7"
djangorestframework = "^3.15.2"
pytest-django = "^4.8.0"
pytest = "^8.3.2"
faker = "^26.1.0"
django-extensions = "^3.2.3"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
django-simple-captcha = "^0.6.0"
bleach = "^6.1.0"
python-decouple = "^3.8"
freezegun = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
